fix: move onboarding reminder to UserService, fix index bug, rename noOfReminderSent to timeRemindedForOnboarding, removed outdated column and other fixes
