--- apps/api/src/tasks/tasks.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common'
-import { OnboardingReminderTask } from './onboarding-reminder.task'
-import { MailModule } from '@/mail/mail.module'
-import { PrismaModule } from '@/prisma/prisma.module'
-
-@Module({
-  imports: [MailModule, PrismaModule],
-  providers: [OnboardingReminderTask]
-})
-export class TasksModule {}
