+++ package.json
@@ -190,7 +190,6 @@
     "clsx": "^2.1.0",
     "conventional-changelog-conventionalcommits": "8.0.0",
     "conventional-changelog-writer": "8.1.0",
-    "date-fns": "^4.1.0",
     "framer-motion": "^11.2.9",
     "million": "^3.0.5",
     "moment": "^2.30.1",
